+++ CHANGELOG.md
@@ -17,10 +17,12 @@ the following form: YYYY.0M.0D.
 ### Added
 
 - Designed kernel and enhanced interfaces for Deck component component
+- Designed abstract class for Deck component
 
 ### Updated
 
 - Changed design to include size, drawMult, ShuffleIn, and scry.
+- Changed design to include DeckSecondary drawMult, scry, and shuffleIn method implementations and SimpleCard name and cost method implementations.
 
 ## [2024.12.30]